Start separating out the upload box from the upload queue
